+++ src/Uri/AddressesUri.php
@@ -4,18 +4,16 @@ declare(strict_types=1);
 
 namespace DigipolisGent\Flanders\BasicRegisters\Uri;
 
-use DigipolisGent\API\Client\Uri\UriInterface;
-
 /**
  * Uri where the addresses (adressen) methods are located.
  */
-class AddressesUri implements UriInterface
+class AddressesUri extends AbstractUriWithFiltersAndPager
 {
 
     /**
      * @inheritDoc
      */
-    public function getUri()
+    protected function getPath(): string
     {
         return 'adressen';
     }